create utils and move logic
